feat: filter offer status
