+++ components/bsx/Offer/OffersUserTable.vue
@@ -2,13 +2,13 @@
   <div>
     <b-select v-model="selectedStatus">
       <option
-        v-for="option in uniqType"
+        v-for="option in getUniqType(offers)"
         :value="option.type"
         :key="option.type">
         {{ option.value }}
       </option>
     </b-select>
-    <b-table :data="updatedOffers">
+    <b-table :data="displayOffers(offers)">
       <b-table-column
         cell-class="is-vcentered is-narrow"
         field="nft.name"
@@ -87,7 +87,6 @@ import MetaTransactionMixin from '~/utils/mixins/metaMixin'
 import SubscribeMixin from '~/utils/mixins/subscribeMixin'
 import { notificationTypes, showNotification } from '~/utils/notification'
 import { tokenIdToRoute } from '~/components/unique/utils'
-import { AllOfferStatusType } from '~/utils/offerStatus'
 
 const components = {
   Identity: () => import('@/components/shared/format/Identity.vue'),
@@ -104,37 +103,6 @@ export default class OffersUserTable extends mixins(
 ) {
   @Prop({ type: Array, default: () => emptyArray<Attribute>() })
   public offers!: Offer[]
-  protected offerStatus: AllOfferStatusType = AllOfferStatusType.ALL
-  protected offersUpdated: Offer[] = []
-
-  get uniqType(): { type: AllOfferStatusType; value: string }[] {
-    const statusSet = new Set(this.offers.map((offer) => offer.status))
-    const singleEventList = Array.from(statusSet).map((type) => ({
-      type: type as AllOfferStatusType,
-      value: AllOfferStatusType[type],
-    }))
-    return [{ type: AllOfferStatusType.ALL, value: 'All' }, ...singleEventList]
-  }
-
-  get updatedOffers() {
-    if (!this.offersUpdated.length) {
-      this.offersUpdated = this.offers
-    }
-    return this.displayOffers(this.offersUpdated)
-  }
-
-  get selectedStatus() {
-    return this.offerStatus
-  }
-
-  set selectedStatus(value: AllOfferStatusType) {
-    if (value === AllOfferStatusType.ALL) {
-      this.offersUpdated = this.offers.concat()
-    } else {
-      this.offersUpdated = this.offers.filter((offer) => offer.status === value)
-    }
-    this.status = value
-  }
 
   public timestampOffer(date) {
     return formatDistanceToNow(new Date(date), { addSuffix: true })
